requirements.txt
aiogram==3.6.0
MazgaDB==1.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
apscheduler==3.10.4
python-dotenv==1.0.1